+++ src/provider/Format/FormatProvider.ts
@@ -14,6 +14,8 @@ import { getMatrixAhk2exeKeep } from './tools/getMatrixAhk2exeKeep';
 import { getMatrixFileBrackets } from './tools/getMatrixFileBrackets';
 import { getMatrixMultLine } from './tools/getMatrixMultLine';
 import { getMatrixTopLabe } from './tools/getMatrixTopLabe';
+import type { TCmdSemantic } from './tools/semanticCmd';
+import { getCmdCll, getSemanticCmdShell } from './tools/semanticCmd';
 import { fn_Warn_thisLineText_WARN } from './TWarnUse';
 import type { TLnStatus } from './wantRefactor/getDeepKeywords';
 import { getDeepKeywords } from './wantRefactor/getDeepKeywords';
@@ -96,6 +98,8 @@ export function FormatCore(
     const matrixBrackets: readonly TBrackets[] = getMatrixFileBrackets(DocStrMap);
     const matrixMultLine: readonly (-999 | 0 | 1)[] = getMatrixMultLine(DocStrMap);
     const matrixAhk2exeKeep: readonly boolean[] = getMatrixAhk2exeKeep(DocStrMap);
+    const cmdSemanticList: readonly TCmdSemantic[] = getSemanticCmdShell(AhkFileData);
+    const matrixCll: readonly (0 | 1)[] = getCmdCll(cmdSemanticList);
     const { mainList, betaList } = getFormatFlag(DocStrMap);
 
     let lnStatus: TLnStatus = {
@@ -133,6 +137,7 @@ export function FormatCore(
                     betaList,
                     DocStrMap,
                     cmdTo1_or_2,
+                    cmdSemanticList,
                 }, AhkTokenLine),
             );
         } else if (line > fmtEnd) {